refactor: extract type query functions from codegen_emit.c (#129)
Move 8 type query/resolution functions into new codegen_types.c/h,
leaving codegen_emit.c as purely emission logic. Clean up header
split by consolidating type-related declarations in codegen_types.h.
